+++ packages/pangea-node-sdk/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Sanitize service support.
+
 ### Fixed
 
 - Event canonicalization of characters like a-circumflex and the replacement